+++ src/serial/serial.hpp
@@ -25,7 +25,7 @@ class Serial {
     Serial();
     bool tryConnectTo(const ComPort& port);  // Trying connect to specified port, return true if sucsesful
     void reset();
-    const void* readData(unsigned* length);
+    const void* readData(unsigned long* length);
     void writeData(const char* data, int length);
 };
